Fixing errors after merge
