+++ Core/src/commonMain/kotlin/com/sdk/growthbook/model/GBValues.kt
@@ -1,5 +1,6 @@
 package com.sdk.growthbook.model
 
+data object GBNull: GBValue()
 data class GBBoolean(val value: Boolean): GBValue()
 data class GBString(val value: String): GBValue()
 data class GBNumber(val value: Number): GBValue()
@@ -10,11 +11,18 @@ data class GBJson(
     private val value: Map<String, GBValue>,
 ): GBValue(), Map<String, GBValue> by value
 
+fun String.toGbString() = GBString(this)
 fun Number.toGbNumber() = GBNumber(this)
 fun Boolean.toGbBoolean() = GBBoolean(this)
 
 sealed class GBValue {
     data object Unknown: GBValue()
+
+    fun isPrimitiveValue(): Boolean =
+        when(this) {
+            is GBNull, is GBBoolean, is GBString, is GBNumber -> true
+            else -> false
+        }
 
     companion object {
         internal fun from(anyObj: Any): GBValue =
