+++ GrowthBook/src/commonMain/kotlin/com/sdk/growthbook/model/GBExperiment.kt
@@ -8,6 +8,7 @@ import com.sdk.growthbook.utils.GBBucketRange
 import com.sdk.growthbook.utils.GBVariationMeta
 import com.sdk.growthbook.utils.RangeSerializer
 import com.sdk.growthbook.utils.GBParentConditionInterface
+import com.sdk.growthbook.kotlinx.serialization.gbSerialize
 import com.sdk.growthbook.serializable_model.SerializableGBExperiment
 import com.sdk.growthbook.serializable_model.SerializableGBExperimentResult
